requirements.txt
streamlit
scikit-learn
pdfminer.six
python-docx
spacy
en-core-web-sm
